Add cast to remove conversion warning

+++ src/civetweb.c
@@ -9776,7 +9776,8 @@ handle_websocket_request(struct mg_connection *conn,
 			do {
 				sep = strchr(protocol, ',');
 				curSubProtocol = protocol;
-				len = sep ? (unsigned long)(sep - protocol) : strlen(protocol);
+				len = sep ? (unsigned long)(sep - protocol)
+				          : (unsigned long)strlen(protocol);
 				while (sep && isspace(*++sep))
 					; // ignore leading whitespaces
 				protocol = sep;
